implémenter déplacement tête serpent. Lier mvtSnake avec autres méthodes. Enlevé méthodes mouvement Mattéo. Corrigé collisionSerpentPomme (on crée un bloc de serpent à la queue du serpent dans mvtSnake). Corrigé bug resetLevel (la liste d'objets ne se réinitialisait pas). Cependant : bug les nouveaux blocs existent mais ne s'affichent pas... Implémenté méthode getTeteSerpent.

+++ src/main/java/graphique/MoteurGraphique.java
@@ -52,7 +52,7 @@ public class MoteurGraphique {
     }
     public void display(LinkedList<Objet> objects){
         for (Objet o : objects){
-            o.setBounds((int) o.getXposition(), (int) o.getYposition(), o.getSizeImageX(), o.getSizeImageY());
+            o.setBounds(o.getXposition(), o.getYposition(), o.getSizeImageX(), o.getSizeImageY());
         }
     }
